// packages/dnb-eufemia/src/extensions/forms/constants/countries.d.ts
export type CountryISO = (typeof countries)[number]['iso'];
export type CountryCdc = (typeof countries)[number]['cdc'];
export type CountryType = {
    cdc: string;
    iso: string;
    i18n: {
        en: string;
        nb: string;
    };
    continent: ContinentType;
    regions?: RegionType;
    name?: string;
};
export type RegionType = Array<'Scandinavia' | 'Nordic'>;
export type ContinentType = 'Asia' | 'Europe' | 'Africa' | 'Oceania' | 'Antarctica' | 'North America' | 'South America' | 'None';
export type CountryLang = keyof CountryType['i18n'];